requirements.txt
streamlit
pandas
numpy
pymongo
python-dotenv
matplotlib
scikit-learn
joblib
rasterio
scipy
